libathemecore/memory: fix sodium hardened memory allocator
I didn't actually have the ability to test this on my development
machine before committing this, so I pushed it and then tried it
on my testnet VPS.

There's only one logic error: Inserting the first item into the
memblocks list succeeds, but then that item is mprotected read-
only and inserting further items fails (because it needs to
modify the pointers in the existing item's node).

While fixing this I had a realisation: We should also protect the
list itself, not just its members. The whole point of the sodium
hardened memory allocator feature is to protect against memory
corruption, but if we're only protecting the items in the list,
and not the list itself, it's almost pointless; so, dynamically
allocate the list, too.

Atheme successfully inits and links with this commit and the
feature enabled. I stress again that this feature is not intended
for use by general end-users; if any of the following occur, the
program is immediately aborted -- any changes made since the last
database commit will be lost:

- under-reads in any dynamic allocation
- under-writes in any dynamic allocation
- over-reads in any dynamic allocation
- over-writes in any dynamic allocation
- free(ptr!=NULL) when ptr was not allocated by us
- realloc(ptr!=NULL, 0) when ptr was not allocated by us
- sodium_malloc() (which is used by scalloc() & smalloc()) failing
- sodium_mprotect_readonly() & sodium_mprotect_readwrite() failing
- attempted corruption of the allocated memblocks list or members

To that effect, I have made the configuration messages more verbose
in this case.