backend/corestorage: save & load modules the database depends upon
This flags modules that install database handlers and password crypto
modules for reloading on startup (if not already loaded).

The former because starting up without the modules in the config file is
going to trigger an "unknown directive" error in corestorage, which
means the user has to figure out which module they just removed uses
that directive; and the latter because removing a crypto module from the
configuration and restarting services is likely to leave people unable
to log in.

If the admin truly wants the module gone, they should unload it from a
running Atheme instance and commit the database. Modules that are marked
non-unloadable cannot be removed this way, but such modules are the ones
that provide the most database functionality and thus have the greatest
potential for massive unwanted data loss.

A truly competent services admin will be able to decide whether editing
the MDEP line (and the directives that the module writes, if any) out of
the database is worth the risk of corrupting it (and besides, they will
have made a backup anyway, right?).

Suggested-by: Janik Kleinhoff <janik@kleinhoff.de> (ilbelkyr)