libathemecore/auth.c: verify_password(): swap blocks around for style
The much more complex logic for dealing with encrypted passwords can
be taken out an entire level of indentation if we swap the conditional
check that decides whether it is or not.

While we're at it, warn if the user's password isn't encrypted, and
put the message strings inside _() for future translation.

[ci skip]