configure: don't require --disable-contrib for --enable-sodium-malloc
The contrib module tree is now using s(c|m|re)alloc() & sfree()
exclusively.

The previous concern was if a contrib module used s(c|m|re)alloc()
but then free() instead of sfree() because the latter is a recent
addition and with the sodium memory allocator enabled this would be
wrong.

[ci skip]