modules/proxyscan/dnsbl.c: replace default: in switch() on enum
The entire point of an enumeration is to indicate to the developers
(by giving them a textual name) and the compiler what values are
expected for a given variable.

Having a default case in a switch statement taking an enum only
makes it very likely that if the enum is extended in the future,
incorrect behaviour would result from the developers failing to
consider how the program will act when it sees a value that was
not expected by the developers.

Fixes a compiler diagnostic regarding a switch statement on an
enumeration which does not cover all of the enumeration's possible
values.