m4/atheme-*.m4: avoid common shell test idiom where possible
The usual practice is to `test "x${foo}" = "xbar"` because ${foo}
may be empty when expanded (if it was unset or set to "") and this
may cause a syntax error depending on your shell and its parameter
expansion logic.

It makes sense to use this idiom when we don't know the possible
values of the variable; but when we control its range of possible
values, it looks nicer to not use it. Plus, it's less confusing to
shell-scripting novices.

Avoid using it where possible, dropping the "x" on both sides.