Add libpasswdqc support to complement libcrack2
This renames modules/nickserv/cracklib to modules/nickserv/pwquality
and adds libpasswdqc <https://www.openwall.com/passwdqc/> support to it.

It also moves the cracklib_dict configuration option from nickserv/main
to this new module. It doesn't make sense to have it elsewhere when
this module is the only thing using it.

Also, the cracklib functionality now requires a dictionary to be set and
present, which reflects the example configuration file documentation.

Finally, it reports the configured dictionary in OperServ INFO.