m4/: library tests: inform that config.log contains failure reason
When asked to explicitly build --with-a-library and detection of
that library's usability fails, inform the user that config.log
contains the compiler output that would explain why it failed (e.g.
implicit declaration of a function because the library is too old
to provide the necessary function).

This is achieved by using the AC_MSG_FAILURE macro instead of the
AC_MSG_ERROR macro. The former wraps around the latter and suggests
to look in config.log for more information.

https://www.gnu.org/software/autoconf/manual/autoconf.html#Printing-Messages