configure: allow user to override CFLAGS optimisation level
If --disable-debugging was given, CFLAGS="-O2" was forced.
If --enable-debugging was given, CFLAGS="-O0 -g" was forced.

In either case, the user cannot specify their own -Ox via the
environment. Now we only set CFLAGS under --disable-debugging
(the default option) if it was unset when configure was invoked.

We also test for (but don't require) -g under --disable-debugging
if only so that occasional production backtraces may be usable.
They are not anywhere near as useful as under -O0 (e.g. variables
are usually all optimised out, line numbers may be incorrect, ...)
but the file names and function names are sometimes still useful.

Fixes: 6a7d3169 ("Build System: Enable CFLAGS="-O2" only if
                  --disable-debugging")