modules/crypto/pbkdf2v2: add HMAC-MD5 support (verify-only)
This is cheap (the necessary scaffolding is already present) and easy,
and will enable the migration of password hashes from more sources.

It will always re-encrypt the password, after verification, with a
stronger digest algorithm, because you cannot configure the module to
use this one.

This is a back-port of commit 7b1204391f5bd25b5409728aac46b2855f7f9433
from the master (v7.3) branch, for forward-compatibility.