Build System: Move all -Wno-foo to the same place
Don't try and keep track of which ones are Clang-specific or not, as GCC will
eventually add support for some of Clang's diagnostic switches anyway, and we
would like to be as consistent as possible; the impossibility of reliably
detecting support for a diagnostic flag (or lack of it) on GCC aside [1].

[1] GCC does not support -Werror=unknown-warning-option and does not throw an
    error if passed an unknown -Wfoo. Furthermore it does not throw an error
    if passed an unknown -Wno-foo either; and doesn't even diagnose the latter
    unless it is already emitting other diagnostics for the same file. It's a
    bloody mess.

This does increase the diagnostic count when building with GCC significantly,
but only because of the amount of unsupported -Wno-foo diagnostic switches. If
you filter out all of the "unrecognized command line option" diagnostics, the
total diagnostic count when compiling under GCC with this change is actually
reduced by a little, as expected.