include/base64.h: add a more precise encoding overhead calculator
This macro should be as precise as possible in its calculation of the
buffer length required. It turns out that it is possible to calculate
an exact value in all circumstances. I am ashamed to say that it took
me the better part of a day to figure this out; my mistake started at
the multiplication before division, like in the macro being replaced,
and went from there...

Also add documentation for the -- what must appear strange -- choice
of numeric constant +2 (this ensures division by 3 rounds up).