Build System: Improve logging messages and library detection structure
Don't discount a library (e.g. libcrypto, libmbedcrypto) just because of
the fact that the first reason we added support for them (to stand in to
the Digest API) is not possible due to the configure test for that
functionality failing.

This still lets the library be used in other ways if more tests exist.

This commit also gives consistent, predictable names to the many feature
enablement macros defined by the build system, and adjusts the rest of
the source tree to use those names; it also adds them to get_conf_opts()
in `libathemecore/conf.c` and adjusts the flags for alphabetical order,
capital letters first, since many of those are affected by ./configure.