libathemecore/memory: add a constant-time memory comparison function
This will use the first found implementation from the following list:

- OpenBSD v4.9+ timingsafe_bcmp(3)
- OpenBSD v5.6+ timingsafe_memcmp(3)
- libsodium's sodium_memcmp()
- libcrypto's CRYPTO_memcmp()
- libnettle's nettle_memeql_sec() (negated return value)

If none of the above are found, it falls back to a naive internal
implementation with a preprocessor diagnostic.

Note that this function is NOT suitable for lexicographic ordering
operations. The only guarantee is that it will return 0 if the two
inputs are equal; all other behaviour is unspecified! Do NOT test
this function's return value against anything other than 0!