libathemecore: ARM mbedTLS: support more algorithms and configuration
The previous RNG code was hardcoded for HMAC-DRBG (with SHA2-256).
Now it will use the first one from this list:

- HMAC-DRBG (with SHA2-512)
- HMAC-DRBG (with SHA2-256)
- CTR-DRBG (with AES)

Also add the Atheme version string as personalisation data when reseeding
too; and abort if being asked for random data before being initialised.

Finally, since ARM mbedTLS allows the user to provide a config file with
a command-line macro declaration, support that too.