Atheme Digest Interface: Add data vector functions
All of the supported digest algorithms are Merkle-Damgard constructions,
which is why we have a universal digest_update() function to work with
them, but having to call digest_update() many times with separate data
buffers can be avoided if we implement a digest_update_vector() function
that takes a vector of input buffers and their lengths. Also add a
digest_oneshot_vector() and digest_oneshot_hmac_vector() too for the
sake of completeness.