Digest API: One-shot functions: Relax pointer aliasing optimisation hints
The restrict qualifier indicates that pointers do not alias eachother, and
enables optimisations based on this assumption; but it is conceivable that
a user of this API may wish to pass the same buffer and/or size value for
the input(s) and the output. Remove the qualifier where appropriate.