modules/saslserv/ecdsa-nist256p-challenge: buffer optimisations
Ever since commit 44d0b84e7f248eaa6c41, the SASL core has ensured that
the data passed to a mechanism is NULL-terminated, should that mechanism
wish to treat the data as a string.

Make use of this property to avoid unnecessarily copying the data to a
NULL-terminated buffer before processing it. The AUTHCOOKIE, EXTERNAL,
PLAIN, and SCRAM-SHA-* mechanisms already do just this.