Build System: Library tests: Simplify variables and conditional checks
Some of these variables are private to the m4 function themselves, and
so they don't need to be declared at all if their sole purpose is to
indicate whether a certain functionality is present or not for later
testing. If the functionality is available, the library must be usable
(for some purpose), and so we don't need to defer AC_DEFINE and friends
to the test for usability.

Likewise we don't need to keep setting variables to No when they already
have that value.