Build System: Move activation of conditional compilation to separate file
This leaves all the logic for setting the conditional compilation variables
in two files: extra.mk.in and m4/atheme-conditional-compilation.m4.

This should make referencing any conditional compilation easier in the
future, and enable other people to more easily wrap their heads around what
to do when needing to use it.