operserv/override: remove
The override functionality offered by this module was insufficiently
transparent to users while providing large opportunities for abuse.

Any legitimate use for this command should be met with a more specific,
more accountable command, such as chanserv/fflags or operserv/mode.

Since the module may cause crashes with commands that do not expect to
be called under the unusual circumstances caused by this module, it is
not suitable even for debugging purposes gated behind allow_taint.

If this removal breaks network staff workflow for you, please consider
whether there is a less invasive, more accountable method of doing what
you need to do. If there could be one, but it isn't implemented, feel
free to report it as a missing feature. If what you need to do can only
be done by arbitrarily impersonating users, reconsider whether you
really need to do it.