modules/crypto/pbkdf2v2: add verify-only SCRAM-MD5 support
This will not enable SASL SCRAM-MD5 logins, but it will enable verification
of password hashes generated by a system or services package that did have
support for that.

The password will still be re-encrypted with a stronger digest algorithm,
as it cannot be set as the default; it is verify-only, like the original
HMAC-MD5 support added in commit 7b1204391f5bd25b5409728aac46b2855f7f9433.