nickserv/enforce: fix renaming sequence in REGAIN
We try to hold the nick and rename its current occupier to GuestXXXXX.
The hold step makes sense (to prevent someone taking the nick between
the two renames) only if it comes before the first rename.

Additionally, setting UF_DOENFORCE on the nick we're about to rename to
doesn't make any sense, so remove that.