libathemecore, modules: remove translation from inappropriate calls
When a user has configured a custom language (nickserv/set_language),
services switches to that language when processing one of their
commands.

This means that logcommand(), slog(), verbose() and the like should
NOT have their strings translated; otherwise, a mischievious user can
set their language to one that will not be widely understood and then
wreak havoc.