modules/saslserv/ecdh-x25519-challenge: set pubkey: better validation
Our base64 codec skips over whitespace when decoding, and a struct
command with a maxparc of 1 can result in the function receiving
public key data with whitespace in it. Encode the result of the
decoding to ensure there is no whitespace before using it to test
against the current value, set it as metadata, and report it back
to the user.