modules/: clean up tabular representations of text
Where data is presented in tabular form with column headers and
dashed lines to separate them from the rows of data, use a
consistent number of dashes (either 32 or, if that is not long
enough to encompass the headers, 64 or even 80). Do not insert
spaces in the line to denote where columns end and begin because
the translated column name is likely to be of a different width.

Also bump the alignment for certain things a little in anticipation
of the translations indeed exceeding the previous limits, and add
notes for translators to the effect of the limits may need changing
if this is still insufficient.