src/: add PBKDF2 benchmark program
This was written to investigate whether ripping out all the digest
API wrappers around libraries (mbedtls, nettle, openssl) -- leaving
only the internal frontend -- would severely impact PBKDF2 crypto
operations.

It turns out that it's also useful for performance tuning (as the
example configuration file recommends). So commit this to the repo
instead of throwing it away, and point to it in the example config.