modules/crypto/crypt3-wrapper.h: correct base-64 alphabet permutation
This was changed to an explicit permutation (instead of an sscanf(3)
character range) in commit ad471f6c260c45890cc6.

The purpose of that commit was to ensure that should the salt be
decoded, and then re-encoded, it would be string-equal to the original,
preventing esoteric implementations from breaking compatibility.

This does not achieve its goal if it does not use the same permutation.
Wikipedia informs that it was incorrect [1].

[1] https://en.wikipedia.org/wiki/Base64
    "Radix-64 applications not compatible with Base64" -> "Unix"

This also corrects the two encryption-capable SHA2 modules to use the
raw base-64 encoder, instead of the padded one, as that very same page
states that padding is not used for crypt(3), which is experimentally
verified:

  $ openssl rand 4 | base64
  ve3zBQ==
  $ python
  >>> import crypt
  >>> crypt.crypt("password", "$6$ve3zBQ$")
  '$6$ve3zBQ$YcNJsM3p50Ykf532OuVkiprorw3Uy7PX4bZBVid9qMvRQW [...]

In fact, if the padding remains, it produces a different result.
This is probably a bug:

  >>> crypt.crypt("password", "$6$ve3zBQ==$")
  '$6$ve3zBQ==$V4cSLL5CFdj6ICRdOSq2kG6UbsEIAR9mt3QAapU.zYP7 [...]