modules/crypto/: add verify-only Anope v2.0+ compatible enc_sha256
Verify-only module for Anope v2.0+ "enc_sha256" password hashes. This is
a straight SHA2-256 invocation but with a custom initial hash value,
which sets this module apart from modules/crypto/rawsha2-256.

It also means that we can't use our Digest API directly, because it will
usually be provided by a crypto library, which do not permit changing
the IHVs. So we use our own internal digest algorithm implementation
directly.

Note that this module differs from the Anope one in significant ways:

1: Anope uses a prefix of "sha256:", while this module uses a prefix of
   "$anope$enc_sha256$".

2: The 2 values (message digest and IHV) are encoded in base-64 instead
   of Anope's hexadecimal.

3: Fields are separated with "$" instead of Anope's ":".

4: The order of the two fields is swapped; the IHV is first.

It is the responsibility of a database migration program to perform
these adaptations.