m4/atheme-libtest-mowgli: default to internal mowgli submodule
The libmowgli packaged on many distributions (if at all) is too
old for the master branch to successfully build against.

This manifests as LGTM (https://lgtm.com/) being unable to build
any pull requests against the master branch, because they hook
our ./configure script to look at what libraries it is testing
for, pause the script there, and then install whatever package
provides that library before resuming (so that the test can
succeed).