modules/saslserv/scram-sha: avoid unregister/register ping-pong
On rehash, if the new mechanism we would choose is the same mechanism we
already registered, don't do anything with regards to unregistering and
registering mechanisms. This saves some unnecessary S2S traffic and
removes a very narrow window of possible confusion wrt a SCRAM mechanism
not being in an IRCv3.2 server's "sasl" CAP value for a newly-connecting
client.