libathemecore/services.c: handle_nickchange(): transparency improvement
Warn user if debugging is enabled regardless of whether it's to a
console or not; greater transparency is always a good thing.

A situation encountered the other day in our support channel (a 25.2 GB
var/atheme.log) could most easily be explained by someone having "debug"
in `serverinfo::loglevel`. This will result in network traffic being
logged to that file. It is an easy condition to overlook, which this
commit will also hopefully help to prevent.

Also take the opportunity to tidy up the code a bit; const-correctness,
wrapped text, declaration and assignment in one step, etc.