m4/atheme-libtest-rt.m4: don't link everything against -lrt
If using a libc which requires -lrt for clock_gettime(2), link only
the PBKDF2 benchmark program against that, instead of all binaries.
This is due to AC_SEARCH_LIBS adding the library (if any) to the
LIBS variable, so save it first and restore it after, like we do
for other library tests.

[ci skip]