m4/atheme-libtest-*.m4: remove unnecessary header checks
It's pointless to invoke AC_CHECK_HEADERS when we're going to subsequently
invoke AC_COMPILE_IFELSE / AC_LINK_IFELSE and include that header -- that
test will also fail if the header is not available.

Also note library tests that are still not using pkg-config(1), and add
the same CPPFLAGS saving/restoring from commit 7bf34bc260826e0da2eb to
those files as well, in case they are migrated in the future (this is
currently a no-op, but an anti-regression measure).