module.h: MODULE_TRY_REQUEST_SYMBOL: unconditionally module_request()
The MODULE_TRY_REQUEST_SYMBOL macro first does a module_locate_symbol()
and only calls module_request() to attempt to load the module if it
failed, before trying again. The reasoning is that it will only fail the
majority of the time due to the module not being loaded yet.

The problem is that module_locate_symbol() logs an error-level message
if the module is not (yet) loaded, which makes it very noisy during a
normal startup if the user doesn't have preceeding loadmodule lines for
all of a module's dependencies.

Switch to unconditionally calling MODULE_TRY_REQUEST_DEPENDENCY from
MODULE_TRY_REQUEST_SYMBOL so that the module is always loaded (if it is
not yet loaded), before calling module_locate_symbol(). This reduces
noise during an ordinary startup for a problem that isn't actually a
problem at all.

[ci skip]