modules/saslserv/scram-sha.c: check client nonce consists of valid chars
I noticed on my testnet that someone is implementing SCRAM in their IRC
client, but accidentally sent a nonce with an ASCII newline in it.

RFC 5802 states that nonces can only consist of 'printable' characters,
and gives the range 0x21-0x2B,0x2D-0x7E (inclusive). Check all of the
characters are in that range.

While we're at it, normalise the various error codes reported for an
unacceptable nonce sent by the client.