doc/SASL-SCRAM-SHA: improve advice on mechanism choice
While mentioning #658 to @aaronmdjones, we found the footnote to be a
bit verbose. Add a quick summary to point (3), and split up the actual
footnote into paragraphs to make it a little easier to read.

I'm also adjusting the part about -512 a little to highlight the
compatibility concerns a bit more.

Closes #658, pointed out by @mquin.