Revert "modules/crypto/argon2: grab parameters from context"
This reverts commit c459a4d0e666b889e159ccc5e2df007c7f158cd9.

If the module does indeed adjust its options, then we write the
adjusted options to the parameter string, which will trigger a
re-encryption after every authentication due to the parsed
parameters not matching the configured ones.

Just rely on the library to get it right every time. This may
affect interoperability with other Argon2 implementations, if
they do not adhere to the reference implementation's behaviour
in this regard.

Reported-by: Doug Freed <dwfreed@mtu.edu>