Build System: Improve library tests
For libraries that use pkg-config to detect their presence, make
pkg-config optional. This allows the user to set LIBFOO_CFLAGS and
LIBFOO_LIBS and still have it work, whether they have pkg-config
installed or not.

Also make it possible to specify the (full) path to the library
installation directory as an argument to the --with-foo= option.
This assumes that there are "include" and "lib" directories under
it, so use this as a last resort; override pkg-config (as described
above) instead if at all possible.