src/crypto-benchmark/optimal.c: cap iterations to starting value
In the extremely unlikely event that the initial iteration count
is faster than the target time, cap the derived iteration count
to the initial one, to prevent it from being exceeded. This
currently only happens on CI (with a reduced initial count) which
has no consequences, but if in some magical pixie dust future
world CPUs become fast enough to do the maximum iteration count
within an acceptable timeframe, don't arrive at a value higher
than the maximum.

[ci skip]