modules/saslserv/scram-sha: rename to scram
It's conceivable in the future that SCRAM would be updated to
include other kinds of password hashing algorithms, such as
scrypt and/or Argon2. The name of the mechanism is itself also
just "SCRAM", so name the module that instead, while we can
rename it now to avoid breaking configuration compatibility.

While we're at it, make the module register all supported
mechanisms. This is because changing the digest algorithm
should not break logins that would have continued to succeed
before doing so. Clients can try each mechanism in turn until
one succeeds, or cache the configuration and credentials that
succeeded last time and continue to re-use them without even
noticing.