Digest API: Remove Nettle frontend
This will never work with '-fsanitize=cfi' or '-fsanitize=function' due
to the necessary function pointer casts to incompatible prototypes.

(nettle_md5_init takes different arguments than nettle_sha1_init etc.)

While we're at it, add some missing error/sanity checking to the digest
API, and denote in the ifdef soup and './configure --help' output the
order in which frontends are searched for.