libathemecore/digest_testsuite: remove PBKDF2 no-pass/no-salt vectors
Some cryptographic libraries don't tolerate being told to perform a
PBKDF2 computation with no password and/or no salt. It is well-defined
behaviour to do so, but it's probably just an anti-footgun measure on
the part of the library authors.

However, nothing in this codebase performs such a PBKDF2 calculation.
So just remove the test vectors and prohibit such behaviour in our
Digest API.