Digest API: Add GNU libgcrypt binding
This makes up for the recent removal of the GNU Nettle binding, which
was necessary for the compiler sanitizers feature (-fsanitize=function).
See the commit message in ed8a3fce8f196d139732 for more information.

This library does not tolerate PBKDF2 with a no-password and/or no-salt
situation, which is why the previous commit was necessary.