modules/saslserv/scram: provide correct mechanism to use on mismatch
When a client tries to login with e.g. SCRAM-SHA-512 but their database
credentials were calculated with SHA2-256, indicate to the client that
SCRAM-SHA-256 is the only acceptable SASL SCRAM method for login. This
is done by recalculating a new SASL mechanism list, against a list of
mechanisms to /avoid/ putting into the list, and then sending that list
to the client (but not telling servers to set their general mechanism
list to it, as that would affect other clients). The list of mechanisms
to avoid is calculated based on the database credentials, which we only
know after receiving a username from the client.

As an aside, I am rather disappointed with the design of SCRAM, in
particular its choice to leave negotiation of digest algorithm out of
band. This would be much cleaner if the mechanism was just named "SCRAM"
(or "SCRAM-PLUS"), and the server could indicate (in the server's first
message) which digest algorithm the client should use to calculate
SaltedPassword, ClientKey, & ServerKey. As it stands, we can only fail
the mechanism if the digest algorithm implied by the SASL mechanism name
doesn't match the digest algorithm used to calculate the database
credentials, and then hope that the client eventually falls back to the
correct mechanism.

This commit aids that fallback process for IRCv3.1 networks (IRCv3.2
networks will have the full mechlist as a value of the sasl= capability,
so clients always know which mechanisms it should try). To that end, we
also put a note in the documentation that deploying SCRAM on IRCv3.1
networks is discouraged.