nickserv/info: Make private times more private
The previous approach rounded the elapsed time down to weeks, which does
reduce the resolution of last seen times to a week over one observation.
Unfortunately, if we're allowed multiple observations, we can zero in on
the time the low-resolution clock ticks over.

The solution used here is to reduce the resolution of the last-seen time
itself. Since the time is not changing, this happens the same way each
time, and subsequent observations reveal no information. This is not
perfect on its own: it discloses high-resolution information about users
who were last seen at the beginning of the current Atheme week, since
they couldn't have been last seen in the future. We deal with this by
representing both 0- and 1-week differences as "less than two
weeks ago".

Finally, this is of limited value as long as we reveal the
high-resolution fact that a user is online now. We dispense with this
and show an obfuscated time for private users regardless of
logged-in status.