m4/atheme-libtest-*.m4: pkg-config: actually no-op on not found
Providing an empty action-if-not-found argument for PKG_CHECK_MODULES
doesn't stop ./configure from aborting. Actually do something (with no
effect) instead; we don't require any of these libraries.

This allows the configure script to succeed on a freshly-installed
machine without the libraries in question installed.

Further, warn the user that pkg-config is missing if it is. This is
necessary for a lot of the library detection tests, unless the user
manually specifies LIBFOO_CFLAGS and LIBFOO_LIBS on the ./configure
command-line, which we don't want to encourage, and is provided only
for compatibility with weird operating systems that install libraries
into paths that its toolchain does not look for libraries in.

[ci skip]