nickserv: Prefer advertising two-argument IDENTIFY
Many users are only ever exposed to the one-argument variant, and
require human assistance to get out of situations that could have been
resolved easily. For example, a user matching a +q $~a may find
themselves stuck on an alternate nick.

Attempt to mitigate this problem by making two-argument IDENTIFY more
prominent in help and nags.