command_exec: avoid crashing with security/cmdperm
Previously, if a custom command_authorize handler denied execution of a
command, the error message would unconditionally try to include the
command's normally required privilege (either from the command
definition or an access{} item in the configuration). In the case of a
command normally available without special privileges (such as
nickserv/help), this would cause a null pointer dereference.

In such a case, we now give a generic "you are not authorized"
message instead.

There are some cases where the error message is misleading – in
particular, it'll blame the failure on the defined privileges, even if
the failure is actually due to a custom security policy module
(including the case where the privileges are actually held by the user).
Improving this situation is out of the scope of this commit.