Inspircd: improvements to the module's cmode stuff
More extban types are added, a lot of mode parameter stuff was cleaned
up, and support for some contrib modules was added too. In addition,
durations are now supported for m_chanhistory.

Still to do: make it work with protocol 1205. But that's outside the
scope of the work I've done.