chanserv/why: don't leak akick reason to users without +A/auspex
On channels with the PUBACL flag, we expose the flag listing to users
without any access as well. However, details about AKICK entries are
still gated behind the +A flag. This specifically includes the reason
(including the private part, if any).

Ensure users have +A or the chan:auspex privilege before revealing the
reason to them via the WHY command.