libathemecore/module.c: improve dependency tracking
MODULE_TRY_REQUEST_DEPENDENCY() invokes module_request(), which, contrary
to the name of the former, does not actually add a dependency. This could
allow a crash if a module depends on another module (without importing any
symbols from that other module) that can be unloaded or reloaded; as the
first module will not be told about it, and so will not get an opportunity
to clean up.

While we're at it, thoroughly document the logic for tracking which module
is being initialised and avoiding dependency loops, and move more of it
into the conditional block that tests if a module has an initialiser, as
that is the only part of a module that is allowed to import other modules
anyway. This allows a small optimisation in the error logging for circular
dependencies.

Reported-by: Nicole Kleinhoff <ilbelkyr@atheme.org> (nicole)